Add SHA-512/t support. Improve testing.
SHA-512/t support is provided by a template function (sha512_t) that
takes the bit length as a template parameter. This implementation
operates on bytes, not bits, so only bit lengths that are multiples of
eight are permitted.

The most common forms of SHA-512/t (SHA-512/224 and SHA-512/256) have
their own functions for performance (as the initial hashes are
precalculated) and convenience.

Testing has been extended to include SHA-512/224 and SHA-512/256,
naturally, while also adding more tests.