Rewrite endianness conversion and type punning
The C++11 standard makes very little assurance that type punning through
unions is not undefined behaviour. Section 9.2/19 states:

If a standard-layout union contains two or more standard-layout structs
that share a common initial sequence, and if the standard-layout union
object currently contains one of these standard-layout structs, it is
permitted to inspect the common initial part of any of them. Two
standard-layout structs share a common initial sequence if corresponding
members have layout-compatible types and either neither member is a
bit-field or both are bit-fields with the same width for a sequence of
one or more initial members.

Aside from this exception, which didn't apply to the code replaced by
this commit, C++11 section 9.5/1 states:

In a union, at most one of the non-static members can be active at any
time, this is, the value of at most one of the non-static data members
can be stored in a union at any time.

This, plus the fewer lines required, is the rationale behind this
commit.

GCC does explicitly support type-punning through unions but this is not
part of the ISO standard.

The current code relies on reinterpret_cast and seems quite likely to be
undefined behaviour also. Future research may lead to changing this code
again.